+++ openhands/server/middleware.py
@@ -1,9 +1,11 @@
+from typing import Callable
 import asyncio
 from collections import defaultdict
 from datetime import datetime, timedelta
 from urllib.parse import urlparse
 
-from fastapi import Request
+import jwt
+from fastapi import Request, APIRouter
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import JSONResponse
 from starlette.middleware.base import BaseHTTPMiddleware
@@ -100,3 +102,78 @@ class RateLimitMiddleware(BaseHTTPMiddleware):
                 headers={'Retry-After': '1'},
             )
         return await call_next(request)
+
+class AttachSessionMiddleware:
+    def __init__(self, app, target_router: APIRouter):
+        self.app = app
+        self.target_router = target_router
+        self.target_paths = {route.path for route in target_router.routes}
+        self.prefix = target_router.prefix or ''
+
+    async def __call__(self, request: Request, call_next: Callable):
+        do_attach = False
+        current_path = request.url.path
+        if current_path.startswith(self.prefix):
+            path_without_prefix = current_path[len(self.prefix) :]
+        else:
+            path_without_prefix = current_path
+        if not path_without_prefix.startswith('/'):
+            path_without_prefix = '/' + path_without_prefix
+        if path_without_prefix in self.target_paths:
+            do_attach = True
+
+        if request.method == 'OPTIONS':
+            do_attach = False
+
+        if not do_attach:
+            return await call_next(request)
+
+        user_verifier = UserVerifier()
+        if user_verifier.is_active():
+            signed_token = request.cookies.get('github_auth')
+            if not signed_token:
+                return JSONResponse(
+                    status_code=status.HTTP_401_UNAUTHORIZED,
+                    content={'error': 'Not authenticated'},
+                )
+            try:
+                jwt.decode(signed_token, config.jwt_secret, algorithms=['HS256'])
+            except Exception as e:
+                logger.warning(f'Invalid token: {e}')
+                return JSONResponse(
+                    status_code=status.HTTP_401_UNAUTHORIZED,
+                    content={'error': 'Invalid token'},
+                )
+
+        if not request.headers.get('Authorization'):
+            logger.warning('Missing Authorization header')
+            return JSONResponse(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                content={'error': 'Missing Authorization header'},
+            )
+
+        auth_token = request.headers.get('Authorization')
+        if 'Bearer' in auth_token:
+            auth_token = auth_token.split('Bearer')[1].strip()
+
+        request.state.sid = get_sid_from_token(auth_token, config.jwt_secret)
+        if request.state.sid == '':
+            logger.warning('Invalid token')
+            return JSONResponse(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                content={'error': 'Invalid token'},
+            )
+
+        request.state.conversation = await session_manager.attach_to_conversation(
+            request.state.sid
+        )
+        if request.state.conversation is None:
+            return JSONResponse(
+                status_code=status.HTTP_404_NOT_FOUND,
+                content={'error': 'Session not found'},
+            )
+        try:
+            response = await call_next(request)
+        finally:
+            await session_manager.detach_from_conversation(request.state.conversation)
+        return response
